pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pepper"
version = "0.1.0"
description = "Pepper — sovereign local-first AI life assistant"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.135,<0.136",
    "uvicorn[standard]>=0.44,<0.45",
    "anthropic>=0.94,<0.95",
    "httpx>=0.28,<0.29",
    "psycopg2-binary>=2.9,<3.0",
    "pgvector>=0.4,<0.5",
    "sqlalchemy[asyncio]>=2.0,<2.1",
    "asyncpg>=0.31,<0.32",
    "sentence-transformers>=5.4,<6.0",
    "apscheduler>=3.11,<4.0",
    "python-telegram-bot>=22.7,<23.0",
    "pydantic-settings>=2.13,<3.0",
    "structlog>=25.5,<26.0",
    "tenacity>=9.1,<10.0",
    # Calendar subsystem
    "google-api-python-client>=2.194,<3.0",
    "google-auth-httplib2>=0.3,<0.4",
    "google-auth-oauthlib>=1.3,<2.0",
    # Slack subsystem
    "slack-sdk>=3.41,<4.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "ruff",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["agent*", "subsystems*"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
